all action servers combined to one
